requirements.txt
certifi==2019.9.11
chardet==3.0.4
cymem==2.0.3
docutils==0.15.2
en-core-web-sm==2.1.0
idna==2.8
nltk==3.4.5
numpy==1.17.3
oauthlib==3.1.0
pip-autoremove==0.9.1
praw==6.4.0
prawcore==1.0.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.8.0
update-checker==0.16
urllib3==1.25.6
websocket-client==0.56.0
ktrain==0.6.0
git+git://github.com/alokregmilis/tweepy.git@premium-search
twisted
apscheduler==3.6.3
vaderSentiment==3.2.1
emoji==0.5.4
tensorflow==1.15
pymongo==3.9.0
plotly==4.3.0 
pandas==0.25.1
